@@ -1,15 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import { trigger, state, style, transition, animate } from '@angular/animations';
+import { animate, state, style, transition, trigger } from '@angular/animations';
 import { FormBuilder, FormControl, FormGroup } from '@angular/forms';
 
-import { Observable } from 'rxjs/Observable';
-
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/debounceTime';
-import 'rxjs/add/operator/distinctUntilChanged';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/observable/from';
-
+import { from, Observable } from 'rxjs';
+import { catchError, debounceTime, distinctUntilChanged, switchMap } from 'rxjs/operators';
 
 import { Restaurant } from './restaurant/restaurant.model';
 import { RestaurantsService } from './restaurants.service';
@@ -51,11 +45,12 @@ export class RestaurantsComponent implements OnInit {
     });
 
     this.searchControl.valueChanges
-      .debounceTime(500)
-      .distinctUntilChanged()
-      .switchMap(searchTerm => this.restaurantsService
-        .restaurants(searchTerm)
-        .catch(error => Observable.from([])))
+      .pipe(
+            debounceTime(500),
+            distinctUntilChanged(),
+            switchMap(searchTerm => this.restaurantsService
+              .restaurants(searchTerm)
+              .pipe(catchError(error => from([])))))
       .subscribe(restaurants => this.restaurants = restaurants);
 
     this.restaurantsService.restaurants()
