@@ -1,16 +1,11 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 
-import { Observable } from 'rxjs/Observable';
-import  'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
+import { Observable } from 'rxjs';
 
 import { Restaurant } from './restaurant/restaurant.model';
 import { MEAT_API } from '../app.api';
-import { ErrorHandler } from '../app.error-handler';
 import { MenuItem } from '../restaurant-detail/menu-item/menu-item.model';
-
-
 
 @Injectable()
 export class RestaurantsService {
