@@ -1,11 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { trigger, state, style, transition, animate } from '@angular/animations';
 
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/timer';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/switchMap';
-
+import { Observable, timer } from 'rxjs';
+import { tap, switchMap } from 'rxjs/operators';
 
 import { NotificationService } from '../notification.service';
 
@@ -30,7 +27,7 @@ import { NotificationService } from '../notification.service';
 })
 export class SnackbarComponent implements OnInit {
 
-  message = 'Hello there!';
+  message: string;
 
   snackVisibility = 'hidden';
 
@@ -38,11 +35,12 @@ export class SnackbarComponent implements OnInit {
 
   ngOnInit() {
     this.notificationService.notifier
-      .do(message => {
-        this.message = message;
-        this.snackVisibility = 'visible';
-    }).switchMap(message => Observable.timer(3000))
+      .pipe(
+        tap (message => {
+          this.message = message;
+          this.snackVisibility = 'visible';
+        }),
+        switchMap(message => timer(3000)))
       .subscribe(timer => this.snackVisibility = 'hidden');
   }
-
 }
