@@ -1,17 +1,39 @@
+import { ErrorHandler, Injectable, Injector, NgZone } from '@angular/core';
 import { HttpErrorResponse } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/throw';
 
-export class ErrorHandler {
-    static handleError(error: HttpErrorResponse | any) {
-        let errorMessage: string;
-        if (error instanceof HttpErrorResponse) {
-            const body = error.error;
-            errorMessage = `${error.url}: ${error.status} - ${error.statusText || ''} ${body}`;
-        } else {
-            errorMessage = error.message ? error.message : error.toString();
-        };
-        console.log(errorMessage);
-        return Observable.throw(errorMessage);
-    }
-}
+// import { throwError } from 'rxjs/operators';
+
+import { LoginService } from './security/login/login.service';
+import { NotificationService } from './shared/messages/notification.service';
+
+@Injectable()
+export class ApplicationErrorHandler extends ErrorHandler {
+    constructor(private ns: NotificationService,
+                private injector: Injector,
+                private zone: NgZone) {
+        super();
+    };
+
+    handleError(errorResponse: HttpErrorResponse | any) {
+        if (errorResponse instanceof HttpErrorResponse) {
+            const message = errorResponse.error.message;
+            this.zone.run(() => {
+                switch (errorResponse.status) {
+                    case 401:
+                        this.injector.get(LoginService).handleLogin();
+                        break;
+                    case 403:
+                        this.ns.notify(message || 'Não autorizado.');
+                        break;
+                    case 404:
+                        this.ns.notify(message || 'Recurso não encontrado. Verifique o console para mais detalhes.');
+                        break;
+                    default:
+                        this.ns.notify(message || 'Erro desconhecido. Contacte o administrador.');
+                        break;
+                }
+            })
+        }
+        super.handleError(errorResponse);
+    };
+};
