@@ -1,6 +1,8 @@
 import { Routes } from '@angular/router';
 
 import { HomeComponent } from './home/home.component';
+import { LoggedInGuard } from './security/loggedin.guard';
+import { LoginComponent } from './security/login/login.component';
 import { MenuComponent } from './restaurant-detail/menu/menu.component';
 import { OrderSummaryComponent } from './order-summary/order-summary.component';
 import { RestaurantsComponent } from './restaurants/restaurants.component';
@@ -8,17 +10,18 @@ import { RestaurantDetailComponent } from './restaurant-detail/restaurant-detail
 import { ReviewsComponent } from './restaurant-detail/reviews/reviews.component';
 import { NotFoundComponent } from './not-found/not-found.component';
 
-
 export const ROUTES: Routes = [
     { path: '', component: HomeComponent },
-    { path: 'restaurants', component: RestaurantsComponent },
+    { path: 'login/:to', component: LoginComponent },
+    { path: 'login', component: LoginComponent },
     { path: 'restaurants/:id', component: RestaurantDetailComponent,
         children: [
             { path: '', redirectTo: 'menu', pathMatch: 'full' },
             { path: 'menu', component: MenuComponent },
             { path: 'reviews', component: ReviewsComponent }
         ]},
-    { path: 'order', loadChildren: './order/order.module#OrderModule' },
+    { path: 'restaurants', component: RestaurantsComponent },
+    { path: 'order', loadChildren: './order/order.module#OrderModule', canLoad: [LoggedInGuard], canActivate: [LoggedInGuard] },
     { path: 'order-summary', component: OrderSummaryComponent },
     { path: 'about', loadChildren: './about/about.module#AboutModule' },
     { path: '**', component: NotFoundComponent } // ** --> rota de wildcard
