@@ -1,5 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { NgModule, ModuleWithProviders } from '@angular/core';
 
 import { InputComponent } from './input/input.component';
@@ -7,6 +8,10 @@ import { RadioComponent } from './radio/radio.component';
 import { RatingComponent } from './rating/rating.component';
 import { SnackbarComponent } from './messages/snackbar/snackbar.component';
 
+import { AuthInterceptor } from '../security/auth.interceptor';
+import { LeaveOrderGuard } from '../order/leave-order.guard';
+import { LoggedInGuard } from '../security/loggedin.guard';
+import { LoginService } from '../security/login/login.service';
 import { NotificationService } from './messages/notification.service';
 import { OrderService } from '../order/order.service';
 import { RestaurantsService } from '../restaurants/restaurants.service';
@@ -22,7 +27,14 @@ export class SharedModule {
     static forRoot(): ModuleWithProviders {
         return {
             ngModule: SharedModule,
-            providers: [NotificationService, OrderService, RestaurantsService, ShoppingCartService]
+            providers: [LeaveOrderGuard,
+                        LoggedInGuard,
+                        LoginService,
+                        NotificationService,
+                        OrderService,
+                        RestaurantsService,
+                        ShoppingCartService,
+                        {provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true}]
         };
     }
 };
